mas fetch

+++ 14_API_REST/lista-tareas.js
@@ -1,9 +1,14 @@
 import { FetchService } from "./fetch-service.js";
+import {  MENSAJES } from "./mensajes.js"
 
 export class ListaTareas {
     constructor() {
         this.nodoListaTareas = document.querySelector('#lista')
+        this.nodoBtnAdd = document.querySelector('#btnAdd')
+        this.nodoNewTarea = document.querySelector('#inTarea')
+        this.nodoBtnAdd.addEventListener('click', this.addTarea.bind(this))
         this.uRL = 'http://localhost:3000/tareas'
+        this.uRL2 = 'http://localhost:3000/tareas2'
         this.aTareas = []
         this.fetchService = new FetchService()
         this.getTareas()
@@ -27,7 +32,7 @@ export class ListaTareas {
         )
         this.nodoListaTareas.innerHTML = html
         this.aNodosChecks = document.querySelectorAll('[name="is-completa"]')
-        this.aNodosBorrar = document.querySelectorAll('.borrar-tarea')
+        this.aNodosBorrar = document.querySelectorAll('.borrar-activo')
         this.aNodosChecks.forEach(
             item => item.addEventListener('change', this.checkTarea.bind(this))
         )
@@ -41,20 +46,78 @@ export class ListaTareas {
             <li>
             <input type="checkbox" name="is-completa" id="check-${data.id}"
                 data-id="${data.id}" ${data.isComplete ? 'checked' : '' }>
-            <span class="nombre-tarea">${data.name}</span>
+            <span class="nombre-tarea ${data.isComplete ? 'hecho' : '' }">
+            ${data.name}</span>
             <span id="borrar-${data.id}" data-id="${data.id}"
-                class="borrar-tarea">🗑️</span>
+                class="borrar-tarea ${data.isComplete ? 'borrar-activo' : 'inactivo' }">🗑️</span>
             </li>
         `
         return htmlView
     }
 
+    addTarea() {
+        if (!this.nodoNewTarea.value) {return}
+        let newTarea = {
+            name: this.nodoNewTarea.value,
+            isComplete: false
+        }
+        let headers = new Headers()
+        headers.append("Content-Type", "application/json");
+        this.fetchService.send(this.uRL, {
+            method: 'POST', 
+            headers : headers,
+            body: JSON.stringify(newTarea)
+        }).then(
+            response => {
+                this.nodoNewTarea.value = ''
+                console.log(response)
+                this.getTareas()
+            },
+            error => console.log(error)
+        )
+        this.fetchService.send(this.uRL2, {//BACK UP en tareas2
+            method: 'POST', 
+            headers : headers,
+            body: JSON.stringify(newTarea)
+        }).then(
+            response => {
+                this.nodoNewTarea.value = ''
+                console.log(response)
+                this.getTareas()
+            },
+            error => console.log(error)
+        )
+
+
+
+    }
+
     checkTarea(oEv) {
         console.log(oEv.target.dataset.id)
+        console.log(oEv.target.checked)
+        let datos = {
+            isComplete : oEv.target.checked
+        }
+        let url = this.uRL + '/' + oEv.target.dataset.id
+        let headers = new Headers()
+        headers.append("Content-Type", "application/json");
+        console.dir(headers)
+        this.fetchService.send(url, {
+                method: 'PATCH', 
+                headers : headers,
+                body: JSON.stringify(datos)
+            }).then( // () => this.getTareas()
+                    response => {
+                    console.log(response)
+                    this.getTareas()
+                },
+                error => console.log(error)
+            )
     }
 
     borrarTarea(oEv) {
         console.log(oEv.target.dataset.id)
+        if (!window.confirm( MENSAJES.listaTareas.confirmacion)) {return}
         // TODO Borar en Servicio Web
         let url = this.uRL + '/' + oEv.target.dataset.id
         this.fetchService.send(url, {method: 'DELETE' })
@@ -66,5 +129,5 @@ export class ListaTareas {
                 error => console.log(error)
             )
     }
-
+ 
 }
